Enabling the scitos services again.
In the beginning of the project the scitos_msgs where hidden in the scitos drivers repository which made it necessary to check for the existence of the package during compile time because the joystick should also work in simulation where there is not scitos_drivers repository.
Once the scitos_msgs where moved to scitos_common, I removed this check from the CMakeLists but forgot to remove the compiler flag from the code. Therefore the rumble_control was always compiled without these service enabled.
This is fixed now.